fix: repair layout collapse labels and vitest aliases
